+++ src/components/index.js
@@ -1,1 +1,7 @@
 export { default as Notices } from './Notices'
+export {
+	default as NotificationsDrawer,
+	NotificationsList,
+	NotificationItem,
+	DismissAllButton
+} from './NotificationsDrawer'